+++ changelog.md
@@ -4,6 +4,8 @@
 * Witches
   * Added Dark Art Witches Feature. When approaching a Dark Art Witch, she'll use a Villager spawn egg and call a lightning bolt on it. 
   * Reduced flee speed
+  * Witches no longer drink strong healing potions by default. They'll only drink it if are below certain health
+  * Invisibility potions are now a Throwing feature
   * Reduced Slow Falling duration when witches used it
 * Enemy animals (such as Hoglins) no longer gain the Animal attacking AI